Jsonrpc-lean fix with rapidjson
JsonRpc Lean defines its own version of RapidJson::SizeType, which is
different from the one used internally. This causes the layout of
rapidjson::Document to be different, which is problematic if rapidjson
is used both through a DLL and the executable using the DLL. Fixed it
by removing the SizeType definition in jsonrpc lean.